requirements.txt
streamlit
fastapi
uvicorn
wikipedia
pdfplumber
sentence-transformers
faiss-cpu
python-multipart
pytesseract
pymupdf
transformers
python-docx

